llm_eval_backend/requirements.txt
openai==1.54.1
google-generativeai==0.8.3
anthropic==0.39.0
pandas==2.2.3
spacy==3.8.2
django==5.1.3
django-cors-headers==4.6.0
django_ninja==1.3.0
# python -m spacy download en_core_web_sm??